backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
alembic==1.14.0
slowapi==0.1.9
python-multipart==0.0.20
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
pdfplumber==0.11.4